dashboard sa prikazom akcija i grafikom profita/gubitka

+++ portfolio_app/Flask-BackEnd/server.py
@@ -1,16 +1,18 @@
 import sys
 import os
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+import time
 
 from flask import Flask, render_template, request, redirect, url_for, session, make_response, jsonify
 from flask_session import Session
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_cors import CORS  # Uvozimo CORS
-
+from sqlalchemy import func, case
 
 from models.user import User
 from models.transaction import Transaction
 from models.stock import Stock
+from models.portfolio import Portfolio
 from config import Config
 from models import db
 
@@ -181,6 +183,26 @@ def change_password():
 
     return jsonify({'passMessage': 'Password changed successfully'}), 200
 
+#@app.route('/get-portfolio-value', methods=['GET'])
+def get_balance(user_id):
+
+    # Izračunavanje vrednosti portfolija
+    balance = db.session.query(
+        func.sum(
+            case(
+                (Transaction.transaction_type == 'buy', -(Transaction.price*Transaction.quantity)),
+                (Transaction.transaction_type == 'sell', Transaction.price*Transaction.quantity),
+                else_=0
+            )
+        )
+    ).filter(Transaction.user_id == user_id).scalar()
+
+    # Ako nema transakcija, postavi vrednost na 0
+    balance = balance if balance is not None else 0
+
+    #return jsonify({'portfolio_value': portfolio_value})
+    return balance
+
 #dodavanje transakcije
 @app.route('/transactions', methods=['POST'])
 def create_transaction():
@@ -209,20 +231,40 @@ def create_transaction():
     symbol_splt, name_splt = stock_name.split(" - ", 1)
 
     stock = Stock.query.filter_by(name=name_splt, user_id=user_id).first()
+    user_portfolio = Portfolio.query.filter_by(user_id=user_id).first() 
 
     if transaction_type == "buy":
-        if stock:
+        if stock and user_portfolio:
+            if user_portfolio.buy_power < price * quantity:
+                return jsonify({"error": "Not enough funds!"}), 400
+            transaction_val = price * quantity
             stock.quantity += quantity
-            stock.total_price += price * quantity
+            stock.total_price += transaction_val
+            if user_portfolio.balance == 0.0:
+                user_portfolio.balance = get_balance(user_id)
+            user_portfolio.balance -= transaction_val
+            user_portfolio.buy_power -= transaction_val
+        elif user_portfolio is None:
+            return jsonify({"error": "Not enough funds!"}), 400
         else:
+            if user_portfolio.buy_power < price * quantity:
+                return jsonify({"error": "Not enough funds!"}), 400
             stock = Stock(user_id=user_id, symbol=symbol_splt.strip(), name=name_splt, total_price=price * quantity, quantity=quantity)
+            if user_portfolio.balance == 0.0:
+                user_portfolio.balance = get_balance(user_id)
+            user_portfolio.balance -= price*quantity
+            user_portfolio.buy_power -= price*quantity
             db.session.add(stock)
     elif transaction_type == "sell":
-        if stock and stock.quantity >= quantity:
+        if stock and stock.quantity >= quantity and user_portfolio:
             stock.quantity -= quantity
             stock.total_price -= price * quantity
+            user_portfolio.balance += price * quantity
+            user_portfolio.buy_power += price * quantity
             if stock.quantity == 0:
                 db.session.delete(stock)  # Brise akciju ako je kolicina 0
+                db.session.query(Transaction).filter(Transaction.user_id == user_id,
+        Transaction.stock_name == stock_name).delete()
         else:
             return jsonify({"error": "Not enough stocks to sell"}), 400
 
@@ -230,6 +272,89 @@ def create_transaction():
     db.session.commit()
 
     return jsonify({"message": "Transaction recorded successfully"}), 201
+
+#rute za dashboard
+@app.route('/get-stocks', methods=['GET'])
+def get_stocks():
+    if 'user_id' not in session:
+        return jsonify({'error': 'Unauthorized'}), 401
+    
+    user_id = session['user_id']
+
+    # Dohvati sve akcije koje korisnik poseduje
+    stocks = Stock.query.filter_by(user_id=user_id).all()
+
+    # Izvuci simbole koje korisnik poseduje
+    owned_symbols = [stock.symbol for stock in stocks]
+
+    # Prosecne cene 'buy' transakcija po simbolu
+    average_prices = db.session.query(
+        func.substring_index(Transaction.stock_name, ' - ', 1).label('symbol'),
+        func.avg(Transaction.price).label('average_price')
+    ).filter(
+        Transaction.user_id == user_id,
+        Transaction.transaction_type == 'buy',
+        func.substring_index(Transaction.stock_name, ' - ', 1).in_(owned_symbols)
+    ).group_by('symbol').all()
+
+    # Konvertuj prosečne cene u dict za lakši pristup
+    average_price_map = {symbol: round(avg_price, 2) for symbol, avg_price in average_prices}
+
+    # Kreiraj konačnu listu sa dodatom prosečnom cenom
+    stocks_data = [
+        {
+            "symbol": stock.symbol,
+            "name": stock.name,
+            "quantity": stock.quantity,
+            "average_price": average_price_map.get(stock.symbol, 0)  # Ako nema kupovina, stavi 0
+        }
+        for stock in stocks
+    ]
+
+    return jsonify(stocks_data)
+
+#dodavanje sredstava
+@app.route('/add-funds', methods=['POST'])
+def add_funds():
+    if 'user_id' not in session:
+        return jsonify({'error': 'Unauthorized'}), 401
+
+    data = request.json
+    user_id = session['user_id']
+    buy_power = float(data.get("buy_power"))
+
+    user_portfolio = Portfolio.query.filter_by(user_id=user_id).first()
+
+    if user_portfolio:
+        user_portfolio.buy_power += buy_power
+    else:
+        user_portfolio = Portfolio(user_id=user_id, buy_power=buy_power)
+        db.session.add(user_portfolio)
+
+    fundsData = { 'buy_power': user_portfolio.buy_power }
+
+    db.session.commit()
+
+    return jsonify(fundsData)
+
+# Ruta za dobavljanje podataka o novcanim sredstvima 
+@app.route('/funds', methods=['GET'])
+def get_funds():
+    if 'user_id' not in session:
+        return jsonify({'error': 'Unauthorized'}), 401
+
+    user_id = session['user_id']
+    
+    user_portfolio = Portfolio.query.filter_by(user_id=user_id).first()
+    
+    if not user_portfolio:
+        return jsonify({'funds': 0.0, 'balance': 0.0})
+
+    return jsonify({
+        "funds": user_portfolio.buy_power,
+        "balance": user_portfolio.balance        
+    }), 200
+
 
 # Kreiranje baze ako ne postoji
 with app.app_context():
